refactor: extract shared computeLaptopTotals, split ReviewStep editors into ReviewEditDialog
